+++ client-course-schedulizer/src/components/Header/Header/Header.scss
@@ -28,13 +28,3 @@
   text-decoration: none;
   color: white;
 }
-
-.import-label,
-.export-label {
-  padding-top: 6px !important;
-  padding-bottom: 6px !important;
-  padding-left: 16px !important;
-  padding-right: 16px !important;
-  color: black !important;
-  cursor: pointer;
-}
